xlstream-eval: extract formula lexer into its own module
